chore!: move `src/simulat/` -> `simulat/`
