+++ simulat/core/scenes/game_scene/game_map/sidebar.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 
 import logging as lg
 
-from src.simulat.core.colors import SimulatPalette
-from src.simulat.core.surfaces.surface import Surface
+from simulat.core.colors import SimulatPalette
+from simulat.core.surfaces.surface import Surface
 
 
 class Sidebar(Surface):
