+++ simulat/core/surfaces/buttons/button.py
@@ -8,9 +8,9 @@ from typing import Callable, Final
 
 import pygame as pg
 
-from src.simulat.core.colors import BasicPalette, SimulatPalette
-from src.simulat.core.game import simulat
-from src.simulat.core.surfaces.surface import Surface
+from simulat.core.colors import BasicPalette, SimulatPalette
+from simulat.core.game import simulat
+from simulat.core.surfaces.surface import Surface
 
 
 class Button(Surface):
